Hey team, I made some changes to the auth-manager.ts file. Here's the breakdown:
- Removed the 'type' property from the Payload interface.

It seems like the 'type' property was doing nothing but causing confusion, just like trying to reason with a compiler on a Monday morning.
Perhaps we could have left a comment indicating why the 'type' property was removed, just to avoid future head scratching or blame games. We don't want the next developer to be as puzzled as a cat in a room full of laser pointers!

Remember, it's always good to leave some breadcrumbs for the next coder who comes along, as they might end up having as hard of a time as a cowboy watching a solar eclipse.

And let's not forget, if removing a property was as easy as deleting lines of code, we'd all have clean, bug-free code, but sadly it's not that simple. It's as tough as convincing a junior developer not to use 'console.log' for debugging.

But hey, I guess the code won't miss the 'type' property as much as we won't miss the weekend when it's Monday again.

Until next time, folks, keep coding and stay sane!

Files:
M  src/auth/auth-manager.ts

Committed using auto-commit: Mon Jul 01 2024 21:16:19 GMT-0700 (Pacific Daylight Time)
Autocommit project: https://www.npmjs.com/package/@dobuki/autocommit
All the lame jokes were generated by AI.